Patch by Tomas Popela  <tpopela@redhat.com>, Mark Lam <mark.lam@apple.com> on 2017-06-01
Reviewed by Mark Lam.

Re-define PutByIdFlags as a int32_t enum explicitly because it is
stored as an int32_t value in UnlinkedInstruction.  This prevents
a bug on 64-bit big endian architectures where the word order is
inverted (when we convert the UnlinkedInstruction into a CodeBlock
Instruction), resulting in the PutByIdFlags value not being stored in
the 32-bit word that the rest of the code expects it to be in.

* bytecode/PutByIdFlags.h:

+++ Source/JavaScriptCore/bytecode/PutByIdFlags.h
@@ -29,7 +29,7 @@
 
 namespace JSC {
 
-enum PutByIdFlags : intptr_t {
+enum PutByIdFlags : int32_t {
     PutByIdNone = 0,
 
     // This flag indicates that the put_by_id is direct. That means that we store the property without
